fastapi-ca/pyproject.toml
[project]
name = "fastapi-ca"
version = "0.1.0"
description = ""
authors = [
    {name = "조현준",email = "aaiss0927@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[all] (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "dependency-injector (>=4.45.0,<5.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
